Разделение сервиса работы с данными на сервис отображения данных и обработки данных
