+++ DOfficeCore/Services/DiaryBoxProvider.cs
@@ -1,9 +1,20 @@
-﻿using DOfficeCore.Services.Interfaces;
+﻿using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System;
+
+using DOfficeCore.Services.Interfaces;
+using DOfficeCore.Models;
 
 namespace DOfficeCore.Services
 {
     class DiaryBoxProvider : IDiaryBoxProvider
     {
+        private readonly IViewCollectionProvider _ViewCollectionProvider;
+
+        public DiaryBoxProvider(IViewCollectionProvider ViewCollectionProvider)
+        {
+            _ViewCollectionProvider = ViewCollectionProvider;
+        }
         /// <summary>
         /// Добавлением строки в дневник с возможностью удаления вместо повторного добавления
         /// </summary>
@@ -16,6 +27,34 @@ namespace DOfficeCore.Services
             if (string.IsNullOrEmpty(Line)) return DiaryBox;
             if (DiaryBox.Contains(Line + " ")) return DiaryBox.Remove(DiaryBox.IndexOf(Line), Line.Length + 1);
             else return DiaryBox + Line + " ";
+        }
+
+
+        /// <summary>
+        /// Создание дневника, по одной случайной строке из каждого раздела определенного диагноза
+        /// </summary>
+        /// <param name="DataCollection">Коллекция элементов базы данных</param>
+        /// <param name="CurrentSection">Выбранная секция базы данных</param>
+        /// <returns>Случайный дневник</returns>
+        public (string, ObservableCollection<Section>) RandomDiary(List<Section> DataCollection, Section CurrentSection)
+        {
+            string result = "";
+            var rnd = new Random();
+
+            var BlockList = _ViewCollectionProvider.BlocksFromDataToView(DataCollection, CurrentSection);
+            var linesOfDiary = new ObservableCollection<Section>();
+            foreach (Section block in BlockList)
+            {
+                var LineList = _ViewCollectionProvider.LinesFromDataToView(DataCollection, block);
+                if (LineList.Count > 0)
+                {
+                    var section = LineList[rnd.Next(LineList.Count)];
+                    linesOfDiary.Add(section);
+                    result += section.Line + " ";
+                }
+            }
+
+            return (result, linesOfDiary);
         }
     }
 }
